added ImageNet models

+++ download_pretrained.py
@@ -76,9 +76,12 @@ if __name__ == "__main__":
 
     if "all" or "imagenet" in args.files:
         #move files to the correct location
-        os.makedirs("./data/datasets/", exist_ok=True)
-        shutil.move("./models/pretrained/ImageNet/imagenet_128-256_labels.bin", "./data/datasets/imagenet_128-256_labels.bin")
-        shutil.move("./models/pretrained/ImageNet/imagenet_128-256_images.bin", "./data//datasets/imagenet_128-256_images.bin")
+        if os.path.exists("./models/pretrained/ImageNet/imagenet_128-256_labels.bin"):
+            os.makedirs("./data/datasets/", exist_ok=True)
+            shutil.move("./models/pretrained/ImageNet/imagenet_128-256_labels.bin", "./data/datasets/imagenet_128-256_labels.bin")
+        if os.path.exists("./models/pretrained/ImageNet/imagenet_128-256_images.bin"):
+            os.makedirs("./data/datasets/", exist_ok=True)
+            shutil.move("./models/pretrained/ImageNet/imagenet_128-256_images.bin", "./data//datasets/imagenet_128-256_images.bin")
 
     print("Download and extraction complete.")
